bagEarth {caret}	R Documentation

Bagged Earth

Description

A bagging wrapper for multivariate adaptive regression splines (MARS) via the earth function

Usage

bagEarth(x, ...)

## Default S3 method:
bagEarth(x, y, weights = NULL, B = 50, summary = mean,
  keepX = TRUE, ...)

## S3 method for class 'formula'
bagEarth(formula, data = NULL, B = 50, summary = mean,
  keepX = TRUE, ..., subset, weights = NULL, na.action = na.omit)

## S3 method for class 'bagEarth'
print(x, ...)

Arguments

x
matrix or data frame of 'x' values for examples.

...
arguments passed to the earth function

y
matrix or data frame of numeric values outcomes.

weights
(case) weights for each example - if missing defaults to 1.

B
the number of bootstrap samples

summary
a function with a single argument specifying how the bagged predictions should be summarized

keepX
a logical: should the original training data be kept?

formula
A formula of the form y ~ x1 + x2 + ...

data
Data frame from which variables specified in 'formula' are preferentially to be taken.

subset
An index vector specifying the cases to be used in the training sample. (NOTE: If given, this argument must be named.)

na.action
A function to specify the action to be taken if 'NA's are found. The default action is for the procedure to fail. An alternative is na.omit, which leads to rejection of cases with missing values on any required variable. (NOTE: If given, this argument must be named.)

Details

The function computes a Earth model for each bootstap sample.

Value

A list with elements

fit
a list of B Earth fits

B
the number of bootstrap samples

call
the function call

x
either NULL or the value of x, depending on the value of keepX

oob
a matrix of performance estimates for each bootstrap sample

Author(s)

Max Kuhn (bagEarth.formula is based on Ripley's nnet.formula)

References

J. Friedman, “Multivariate Adaptive Regression Splines” (with discussion) (1991). Annals of Statistics, 19/1, 1-141.

See Also

earth, predict.bagEarth

Examples

## Not run: 
library(mda)
library(earth)
data(trees)
fit1 <- earth(x = trees[,-3], y = trees[,3])
set.seed(2189)
fit2 <- bagEarth(x = trees[,-3], y = trees[,3], B = 10)

## End(Not run)

[Package caret version 6.0-79 Index]