caretSBF {caret}	R Documentation

Selection By Filtering (SBF) Helper Functions

Description

Ancillary functions for univariate feature selection

Usage

caretSBF

anovaScores(x, y)

gamScores(x, y)

Arguments

x
a matrix or data frame of numeric predictors

y
a numeric or factor vector of outcomes

Format

An object of class list of length 5.

Details

More details on these functions can be found at http://topepo.github.io/caret/feature-selection-using-univariate-filters.html.

This page documents the functions that are used in selection by filtering (SBF). The functions described here are passed to the algorithm via the functions argument of sbfControl.

See sbfControl for details on how these functions should be defined.

anovaScores and gamScores are two examples of univariate filtering functions. anovaScores fits a simple linear model between a single feature and the outcome, then the p-value for the whole model F-test is returned. gamScores fits a generalized additive model between a single predictor and the outcome using a smoothing spline basis function. A p-value is generated using the whole model test from summary.Gam and is returned.

If a particular model fails for lm or gam, a p-value of 1 is returned.

Author(s)

Max Kuhn

See Also

sbfControl, sbf, summary.Gam

[Package caret version 6.0-79 Index]