dhfr {caret}	R Documentation

Dihydrofolate Reductase Inhibitors Data

Description

Sutherland and Weaver (2004) discuss QSAR models for dihydrofolate reductase (DHFR) inhibition. This data set contains values for 325 compounds. For each compound, 228 molecular descriptors have been calculated. Additionally, each samples is designated as "active" or "inactive".

Details

The data frame dhfr contains a column called Y with the outcome classification. The remainder of the columns are molecular descriptor values.

Value

dhfr
data frame of chemical descriptors and the activity values

Source

Sutherland, J.J. and Weaver, D.F. (2004). Three-dimensional quantitative structure-activity and structure-selectivity relationships of dihydrofolate reductase inhibitors, Journal of Computer-Aided Molecular Design, Vol. 18, pg. 309–331.

[Package caret version 6.0-79 Index]