dotplot.diff.resamples {caret}	R Documentation

Lattice Functions for Visualizing Resampling Differences

Description

Lattice functions for visualizing resampling result differences between models

Usage

## S3 method for class 'diff.resamples'
dotplot(x, data = NULL, metric = x$metric[1], ...)

Arguments

x
an object generated by diff.resamples

data
Not used

metric
a character string for which metrics to plot. Note: dotplot and levelplot require exactly two models whereas the other methods can plot more than two.

...
further arguments to pass to either densityplot, dotplot or levelplot

Details

densityplot and bwplot display univariate visualizations of the resampling distributions. levelplot displays the matrix of pair-wide comparisons. dotplot shows the differences along with their associated confidence intervals.

Value

a lattice object

Author(s)

Max Kuhn

See Also

resamples, diff.resamples, bwplot, densityplot, xyplot, splom

Examples

## Not run: 
#load(url("http://topepo.github.io/caret/exampleModels.RData"))

resamps <- resamples(list(CART = rpartFit,
                          CondInfTree = ctreeFit,
                          MARS = earthFit))
difs <- diff(resamps)

dotplot(difs)

densityplot(difs,
            metric = "RMSE",
            auto.key = TRUE,
            pch = "|")

bwplot(difs,
       metric = "RMSE")

levelplot(difs, what = "differences")

## End(Not run)

[Package caret version 6.0-79 Index]