featurePlot {caret}	R Documentation

Wrapper for Lattice Plotting of Predictor Variables

Description

A shortcut to produce lattice graphs

Usage

featurePlot(x, y, plot = if (is.factor(y)) "strip" else "scatter",
  labels = c("Feature", ""), ...)

Arguments

x
a matrix or data frame of continuous feature/probe/spectra data.

y
a factor indicating class membership.

plot
the type of plot. For classification: box, strip, density, pairs or ellipse. For regression, pairs or scatter

labels
a bad attempt at pre-defined axis labels

...
options passed to lattice calls.

Details

This function “stacks” data to get it into a form compatible with lattice and creates the plots

Value

An object of class “trellis”. The ‘update’ method can be used to update components of the object and the ‘print’ method (usually called by default) will plot it on an appropriate plotting device.

Author(s)

Max Kuhn

Examples

x <- matrix(rnorm(50*5),ncol=5)
y <- factor(rep(c("A", "B"),  25))

trellis.par.set(theme = col.whitebg(), warn = FALSE)
featurePlot(x, y, "ellipse")
featurePlot(x, y, "strip", jitter = TRUE)
featurePlot(x, y, "box")
featurePlot(x, y, "pairs")

[Package caret version 6.0-79 Index]