panel.needle {caret}	R Documentation

Needle Plot Lattice Panel

Description

A variation of panel.dotplot that plots horizontal lines from zero to the data point.

Usage

panel.needle(x, y, horizontal = TRUE, pch = if (is.null(groups))
  dot.symbol$pch else sup.symbol$pch, col = if (is.null(groups))
  dot.symbol$col else sup.symbol$col, lty = dot.line$lty,
  lwd = dot.line$lwd, col.line = dot.line$col, levels.fos = NULL,
  groups = NULL, ...)

Arguments

x, y
variables to be plotted in the panel. Typically y is the 'factor'

horizontal
logical. If FALSE, the plot is ‘transposed’ in the sense that the behaviours of x and y are switched. x is now the ‘factor’. Interpretation of other arguments change accordingly. See documentation of bwplot for a fuller explanation.

pch, col, lty, lwd, col.line
graphical parameters

levels.fos
locations where reference lines will be drawn

groups
grouping variable (affects graphical parameters)

...
extra parameters, passed to panel.xyplot which is responsible for drawing the foreground points (panel.dotplot only draws the background reference lines).

Details

Creates (possibly grouped) needleplot of x against y or vice versa

Author(s)

Max Kuhn, based on panel.dotplot by Deepayan Sarkar

See Also

dotplot

[Package caret version 6.0-79 Index]