plot.gafs {caret}	R Documentation

Plot Method for the gafs and safs Classes

Description

Plot the performance values versus search iteration

Usage

## S3 method for class 'gafs'
plot(x, metric = x$control$metric["external"],
  estimate = c("internal", "external"), output = "ggplot", ...)

Arguments

x
an object of class gafs or safs

metric
the measure of performance to plot (e.g. RMSE, accuracy, etc)

estimate
the type of estimate: either "internal" or "external"

output
either "data", "ggplot" or "lattice"

...
options passed to xyplot

Details

The mean (averaged over the resamples) is plotted against the search iteration using a scatter plot.

When output = "data", the unaveraged data are returned with columns for all the performance metrics and the resample indicator.

Value

Either a data frame, ggplot object or lattice object

Author(s)

Max Kuhn

See Also

gafs, safs, ggplot, xyplot

Examples

## Not run: 
set.seed(1)
train_data <- twoClassSim(100, noiseVars = 10)
test_data  <- twoClassSim(10,  noiseVars = 10)

## A short example
ctrl <- safsControl(functions = rfSA,
                    method = "cv",
                    number = 3)

rf_search <- safs(x = train_data[, -ncol(train_data)],
                  y = train_data$Class,
                  iters = 50,
                  safsControl = ctrl)

plot(rf_search)
plot(rf_search,
	 output = "lattice",
	 auto.key = list(columns = 2))

plot_data <- plot(rf_search, output = "data")
summary(plot_data)
    
## End(Not run)

[Package caret version 6.0-79 Index]