ggplot.rfe {caret}	R Documentation

Plot RFE Performance Profiles

Description

These functions plot the resampling results for the candidate subset sizes evaluated during the recursive feature elimination (RFE) process

Usage

## S3 method for class 'rfe'
ggplot(data = NULL, mapping = NULL, metric = data$metric[1],
  output = "layered", ..., environment = NULL)

## S3 method for class 'rfe'
plot(x, metric = x$metric, ...)

Arguments

data
an object of class rfe.

mapping, environment
unused arguments to make consistent with ggplot2 generic method

metric
What measure of performance to plot. Examples of possible values are "RMSE", "Rsquared", "Accuracy" or "Kappa". Other values can be used depending on what metrics have been calculated.

output
either "data", "ggplot" or "layered". The first returns a data frame while the second returns a simple ggplot object with no layers. The third value returns a plot with a set of layers.

...
plot only: specifications to be passed to xyplot. The function automatically sets some arguments (e.g. axis labels) but passing in values here will over-ride the defaults.

x
an object of class rfe.

Details

These plots show the average performance versus the subset sizes.

Value

a lattice or ggplot object

Author(s)

Max Kuhn

References

Kuhn (2008), “Building Predictive Models in R Using the caret” (http://www.jstatsoft.org/article/view/v028i05/v28i05.pdf)

See Also

rfe, xyplot, ggplot

Examples

## Not run: 
data(BloodBrain)

x <- scale(bbbDescr[,-nearZeroVar(bbbDescr)])
x <- x[, -findCorrelation(cor(x), .8)]
x <- as.data.frame(x)

set.seed(1)
lmProfile <- rfe(x, logBBB,
                 sizes = c(2:25, 30, 35, 40, 45, 50, 55, 60, 65),
                 rfeControl = rfeControl(functions = lmFuncs,
                                         number = 200))
plot(lmProfile)
plot(lmProfile, metric = "Rsquared")
ggplot(lmProfile)

## End(Not run)

[Package caret version 6.0-79 Index]