predict.knnreg {caret}	R Documentation

Predictions from k-Nearest Neighbors Regression Model

Description

Predict the outcome of a new observation based on k-NN.

Usage

## S3 method for class 'knnreg'
predict(object, newdata, ...)

Arguments

object
object of class knnreg.

newdata
a data frame or matrix of new observations.

...
additional arguments.

Details

This function is a method for the generic function predict for class knnreg. For the details see knnreg. This is essentially a copy of predict.ipredknn.

Value

a numeric vector

Author(s)

Max Kuhn, Chris Keefer, adapted from knn and predict.ipredknn

[Package caret version 6.0-79 Index]