predictors {caret}	R Documentation

List predictors used in the model

Description

This class uses a model fit to determine which predictors were used in the final model.

Usage

predictors(x, ...)

Arguments

x
a model object, list or terms

...
not currently used

Details

For randomForest, cforest, ctree, rpart, ipredbagg, bagging, earth, fda, pamr.train, superpc.train, bagEarth and bagFDA, an attempt was made to report the predictors that were actually used in the final model.

The predictors function can be called on the model object (as opposed to the train) object) and the package will try to find the appropriate coed (if it exists).

In cases where the predictors cannot be determined, NA is returned. For example, nnet may return missing values from predictors.

Value

a character string of predictors or NA.

[Package caret version 6.0-79 Index]