print.train {caret}	R Documentation

Print Method for the train Class

Description

Print the results of a train object.

Usage

## S3 method for class 'train'
print(x, printCall = FALSE, details = FALSE,
  selectCol = FALSE, showSD = FALSE, ...)

Arguments

x
an object of class train.

printCall
a logical to print the call at the top of the output

details
a logical to show print or summary methods for the final model. In some cases (such as gbm, knn, lvq, naive Bayes and bagged tree models), no information will be printed even if details = TRUE

selectCol
a logical whether to add a column with a star next to the selected parameters

showSD
a logical whether to show the standard deviation of the resampling results within parentheses (e.g. "4.24 (0.493)")

...
options passed to format

Details

The table of complexity parameters used, their resampled performance and a flag for which rows are optimal.

Value

A matrix with the complexity parameters and performance (invisibly).

Author(s)

Max Kuhn

See Also

train

Examples

## Not run: 
data(iris)
TrainData <- iris[,1:4]
TrainClasses <- iris[,5]

options(digits = 3)

library(klaR)
rdaFit <- train(TrainData, TrainClasses, method = "rda",
                control = trainControl(method = "cv"))
rdaFit
print(rdaFit, showSD = TRUE)

## End(Not run)

[Package caret version 6.0-79 Index]