resampleHist {caret}	R Documentation

Plot the resampling distribution of the model statistics

Description

Create a lattice histogram or densityplot from the resampled outcomes from a train object.

Usage

resampleHist(object, type = "density", ...)

Arguments

object
an object resulting form a call to train

type
a character string. Either "hist" or "density"

...
options to pass to histogram or densityplot

Details

All the metrics from the object are plotted, but only for the final model. For more comprehensive plots functions, see histogram.train, densityplot.train, xyplot.train, stripplot.train.

For the plot to be made, the returnResamp argument in trainControl should be either "final" or "all".

Value

a object of class trellis

Author(s)

Max Kuhn

See Also

train, histogram, densityplot, histogram.train, densityplot.train, xyplot.train, stripplot.train

Examples

## Not run: 
data(iris)
TrainData <- iris[,1:4]
TrainClasses <- iris[,5]

knnFit <- train(TrainData, TrainClasses, "knn")

resampleHist(knnFit)

## End(Not run)

[Package caret version 6.0-79 Index]