scat {caret}	R Documentation

Morphometric Data on Scat

Description

Reid (2015) collected data on animal feses in coastal California. The data consist of DNA verified species designations as well as fields related to the time and place of the collection and the scat itself. The data frame scat_orig contains while scat contains data on the three main species.

Value

scat_orig
the entire data set in the Supplemental Materials

scat
data on the three main species

Source

Reid, R. E. B. (2015). A morphometric modeling approach to distinguishing among bobcat, coyote and gray fox scats. Wildlife Biology, 21(5), 254-262

[Package caret version 6.0-79 Index]