summary.bagEarth {caret}	R Documentation

Summarize a bagged earth or FDA fit

Description

The function shows a summary of the results from a bagged earth model

Usage

## S3 method for class 'bagEarth'
summary(object, ...)

## S3 method for class 'bagFDA'
summary(object, ...)

Arguments

object
an object of class "bagEarth" or "bagFDA"

...
optional arguments (not used)

Details

The out-of-bag statistics are summarized, as well as the distribution of the number of model terms and number of variables used across all the bootstrap samples.

Value

a list with elements

modelInfo
a matrix with the number of model terms and variables used

oobStat
a summary of the out-of-bag statistics

bmarsCall
the original call to bagEarth

Author(s)

Max Kuhn

Examples

## Not run: 
data(trees)
set.seed(9655)
fit <- bagEarth(trees[,-3], trees[3])
summary(fit)

## End(Not run)

[Package caret version 6.0-79 Index]