varImp.gafs {caret}	R Documentation

Variable importances for GAs and SAs

Description

Variable importance scores for safs and gafs objects.

Usage

## S3 method for class 'gafs'
varImp(object, metric = object$control$metric["external"],
  maximize = object$control$maximize["external"], ...)

Arguments

object
an safs or gafs object

metric
a metric to compute importance (see Details below)

maximize
are larger values of the metric better?

...
not currently uses

Details

A crude measure of importance is computed for thee two search procedures. At the end of a search process, the difference in the fitness values is computed for models with and without each feature (based on the search history). If a predictor has at least two subsets that include and did not include the predictor, a t-statistic is computed (otherwise a value of NA is assigned to the predictor).

This computation is done separately for each resample and the t-statistics are averaged (NA values are ignored) and this average is reported as the importance. If the fitness value should be minimized, the negative value of the t-statistic is used in the average.

As such, the importance score reflects the standardized increase in fitness that occurs when the predict is included in the subset. Values near zero (or negative) indicate that the predictor may not be important to the model.

Value

a data frame where the rownames are the predictor names and the column is the average t-statistic

Author(s)

Max Kuhn

See Also

safs, gafs

[Package caret version 6.0-79 Index]